+++ client/src/io/worxbend/gitea4s/http/CombinedStatusParams.scala
@@ -1,0 +1,9 @@
+package io.worxbend.gitea4s.http
+
+final case class CombinedStatusParams(
+    page: Option[Int] = None,
+    limit: Option[Int] = None
+)
+
+object CombinedStatusParams:
+  val default: CombinedStatusParams = CombinedStatusParams()