+++ client/src/io/worxbend/gitea4s/http/PullRequestUpdateStyle.scala
@@ -1,0 +1,5 @@
+package io.worxbend.gitea4s.http
+
+enum PullRequestUpdateStyle(val queryValue: String):
+  case Merge extends PullRequestUpdateStyle("merge")
+  case Rebase extends PullRequestUpdateStyle("rebase")