+++ core/src/io/worxbend/gitea4s/model/GiteaModels.scala
@@ -491,6 +491,20 @@ final case class PullReviewRequestOptions(
 object PullReviewRequestOptions:
   given JsonCodec[PullReviewRequestOptions] = DeriveJsonCodec.gen[PullReviewRequestOptions]
 
+final case class MergePullRequestOption(
+    @jsonField("Do") mergeMethod: MergePullRequestMethod,
+    @jsonField("MergeCommitID") mergeCommitId: Option[String] = None,
+    @jsonField("MergeMessageField") mergeMessageField: Option[String] = None,
+    @jsonField("MergeTitleField") mergeTitleField: Option[String] = None,
+    @jsonField("delete_branch_after_merge") deleteBranchAfterMerge: Option[Boolean] = None,
+    @jsonField("force_merge") forceMerge: Option[Boolean] = None,
+    @jsonField("head_commit_id") headCommitId: Option[String] = None,
+    @jsonField("merge_when_checks_succeed") mergeWhenChecksSucceed: Option[Boolean] = None
+)
+
+object MergePullRequestOption:
+  given JsonCodec[MergePullRequestOption] = DeriveJsonCodec.gen[MergePullRequestOption]
+
 final case class ChangedFile(
     additions: Option[Long] = None,
     changes: Option[Long] = None,